requirements.txt
pandas
scikit-learn
nltk
joblib
streamlit
datasets
sentence-transformers
google-generativeai
tqdm
streamlit
python-dotenv
requests